+++ Assets/Plugins/SouthPointe/Serialization/MessagePack/Attributes/NonSerializedAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace SouthPointe.Serialization.MessagePack
+{
+	[AttributeUsage(AttributeTargets.Field, Inherited = false)]
+	public sealed class NonSerializedAttribute : Attribute
+	{
+		
+	}
+}