+++ Assets/Plugins/SouthPointe/Serialization/MessagePack/Attributes/OnSerializingAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace SouthPointe.Serialization.MessagePack
+{
+	[AttributeUsage(AttributeTargets.Method, Inherited = false)]
+	public sealed class OnSerializingAttribute : Attribute
+	{
+
+	}
+}