+++ Assets/Tests/Benchmark.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Runtime.Serialization;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace SouthPointe.Serialization.MessagePack
 {