+++ Assets/Tests/Editor/SerializationAttributeTest.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Runtime.Serialization;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using UnityEngine;
 
 namespace SouthPointe.Serialization.MessagePack.Tests